Changed ShortPost to extend more general Post Object. Added BeachGroup and had extend from Group Object.
