--- .idea/misc.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="DesignSurface">
-    <option name="filePathToZoomLevelMap">
-      <map>
-        <entry key="..\:/Users/christianamarc/Documents/AndroidStudioProjects/SurfStop/app/src/main/res/drawable/button_selector.xml" value="0.412" />
-        <entry key="..\:/Users/christianamarc/Documents/AndroidStudioProjects/SurfStop/app/src/main/res/drawable/customborder.xml" value="0.3155" />
-        <entry key="..\:/Users/christianamarc/Documents/AndroidStudioProjects/SurfStop/app/src/main/res/layout/activity_login.xml" value="0.3" />
-        <entry key="..\:/Users/christianamarc/Documents/AndroidStudioProjects/SurfStop/app/src/main/res/layout/activity_main.xml" value="0.25" />
-        <entry key="..\:/Users/christianamarc/Documents/AndroidStudioProjects/SurfStop/app/src/main/res/layout/activity_signup.xml" value="0.25" />
-        <entry key="..\:/Users/christianamarc/Documents/AndroidStudioProjects/SurfStop/app/src/main/res/layout/fragment_profile.xml" value="0.390625" />
-        <entry key="..\:/Users/christianamarc/Documents/AndroidStudioProjects/SurfStop/app/src/main/res/layout/fragment_temp_feed.xml" value="0.25" />
-        <entry key="..\:/Users/christianamarc/Documents/AndroidStudioProjects/SurfStop/app/src/main/res/layout/item_shortpost.xml" value="0.3333333333333333" />
-        <entry key="..\:/Users/christianamarc/Documents/AndroidStudioProjects/SurfStop/app/src/main/res/menu/menu_bottom_navigation.xml" value="0.5541666666666667" />
-      </map>
-    </option>
-  </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
-    <output url="file://$PROJECT_DIR$/build/classes" />
-  </component>
-  <component name="ProjectType">
-    <option name="id" value="Android" />
-  </component>
-</project>
